Ptw/attempt nix (#2)
And remove dev containers. Really, my main issue is one of package management. Dev containers provide a way to manage packages by baking everything into a container image. However, they suffer in the performance department on Mac (due to virtualization costs).

Nix seems very powerful and complicated and complex, but I really only want to use it for its package management prowess. This way, I get a reproducible (hopefully), locally installed (definitely) environment that works with my editor.

Except, it doesn't quite work with my tools. I get around this by adding an alias to the developer shell that `nix develop` provides to start VSCode in the terminal in which this starts. That assures VSCode that everything is where it expects. (Previously, I hardcoded VSCode's `dotnet` paths to the nix store, which isn't shareable across machines and seems very brittle.)

I think my general approach with using nix will be limited to package management. I do not want to dispatch to it for building, installing dependencies, etc. For example, if I wanted to apply this to a rust codebase, I would install `rustup` in one of these shells and that would be it.

* First draft

* Update url

I would think that `nixpkgs-22.05` would work, though!

* Remove dev container

* Add code alias